* Allow arbitrary listing base directories
Unfortunetally this breaks themes which attempt to check whether a file exists
or is of a certain type if the base directory does not match the current
working directory. Each entry in $dirArray now provides the 'real_path' item
which can be used if the actual file path of an entry is required.