Allow to download directories as zip files
The feature adds a button to download all files in current directory as zip archive.
For this function to work it is required to have 'zip' utility available on the server.
To enable this function set the 'zip_dirs' option to 'true' in configuration file (it is disabled by default).

By default the zip is created 'on-the-fly' and streamed to client, without any temporary file. This behaviour can be changed by setting 'zip_stream' configuration option to false.

It is also possible to choose compression level for zip files - see 'zip_compression_level' configuration option. The default value is zero for perfomance reasons.
The 'zip_disable' option allows to disable zip downloading for particular set of directories. It is works much the same as 'hidden_files' option.

TODO: When zip is created the script tries to exclude files with respect to 'hidden files' option. But it doesn't do it very well - for example, if 'hidden_files' have 'resources/*' entry, then it will exclude the contents of 'resources' directory on any subdirectory, not just at the root level.